test(server): preserve stable HTTP test isolation
